+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+- Support collections of dynamic assets in dynamic asset files (resolves [#78](https://github.com/NiklasEi/bevy_asset_loader/issues/78))
+  - Remove `Files` in preference of collections of `File`
+    - Previously:
+        ```ron
+        ({
+            "images": Files (
+                paths: ["images/tree.png", "images/player.png"],
+            )
+        })
+        ```
+    - Now:
+        ```ron
+        ({
+            "images": [
+              File(path: "images/tree.png"),
+              File(path: "images/player.png")
+            ]
+        })
+        ```
+
 ## v0.17.0
 - update to Bevy 0.11
 - Use "/" in paths used as keys for mapped collections on all platforms (resolves [#135](https://github.com/NiklasEi/bevy_asset_loader/issues/135))